// src/security/cases/ediscoveryCases/item/operations/item/caseOperationItemRequestBuilderDeleteRequestConfiguration.ts
import { type RequestOption } from '@microsoft/kiota-abstractions';

export interface CaseOperationItemRequestBuilderDeleteRequestConfiguration {
    /**
     * Request headers
     */
    headers?: Record<string, string[]> | undefined;
    /**
     * Request options
     */
    options?: RequestOption[] | undefined;
}